test/inspector/runtime/set-max-call-stack-size-expected.txt

Checks Runtime.setMaxCallStackSizeToCapture.

Running test: testConsoleTrace
Test with max size 0.
{
    args : [
        [0] : {
            type : string
            value : Nested call.
        }
    ]
    executionContextId : <executionContextId>
    timestamp : <timestamp>
    type : trace
}
Test with max size 1.
{
    args : [
        [0] : {
            type : string
            value : Nested call.
        }
    ]
    executionContextId : <executionContextId>
    stackTrace : {
        callFrames : [
            [0] : {
                columnNumber : 12
                functionName : bar
                lineNumber : 3
                scriptId : <scriptId>
                url : test.js
            }
        ]
        parent : {
            callFrames : [
                [0] : {
                    columnNumber : 4
                    functionName : executor
                    lineNumber : 12
                    scriptId : <scriptId>
                    url : test.js
                }
            ]
            description : setTimeout
        }
    }
    timestamp : <timestamp>
    type : trace
}
Test with max size 2.
{
    args : [
        [0] : {
            type : string
            value : Nested call.
        }
    ]
    executionContextId : <executionContextId>
    stackTrace : {
        callFrames : [
            [0] : {
                columnNumber : 12
                functionName : bar
                lineNumber : 3
                scriptId : <scriptId>
                url : test.js
            }
            [1] : {
                columnNumber : 4
                functionName : foo
                lineNumber : 8
                scriptId : <scriptId>
                url : test.js
            }
        ]
        parent : {
            callFrames : [
                [0] : {
                    columnNumber : 4
                    functionName : executor
                    lineNumber : 12
                    scriptId : <scriptId>
                    url : test.js
                }
                [1] : {
                    columnNumber : 9
                    functionName : testConsoleTrace
                    lineNumber : 11
                    scriptId : <scriptId>
                    url : test.js
                }
            ]
            description : setTimeout
        }
    }
    timestamp : <timestamp>
    type : trace
}

Running test: testException
{
    args : [
        [0] : {
            type : string
            value : Nested call.
        }
    ]
    executionContextId : <executionContextId>
    timestamp : <timestamp>
    type : trace
}
{
    args : [
        [0] : {
            type : string
            value : Nested call.
        }
    ]
    executionContextId : <executionContextId>
    stackTrace : {
        callFrames : [
            [0] : {
                columnNumber : 12
                functionName : bar
                lineNumber : 3
                scriptId : <scriptId>
                url : test.js
            }
        ]
        parent : {
            callFrames : [
                [0] : {
                    columnNumber : 4
                    functionName : executor
                    lineNumber : 12
                    scriptId : <scriptId>
                    url : test.js
                }
            ]
            description : setTimeout
        }
    }
    timestamp : <timestamp>
    type : trace
}
{
    args : [
        [0] : {
            type : string
            value : Nested call.
        }
    ]
    executionContextId : <executionContextId>
    stackTrace : {
        callFrames : [
            [0] : {
                columnNumber : 12
                functionName : bar
                lineNumber : 3
                scriptId : <scriptId>
                url : test.js
            }
            [1] : {
                columnNumber : 4
                functionName : foo
                lineNumber : 8
                scriptId : <scriptId>
                url : test.js
            }
        ]
        parent : {
            callFrames : [
                [0] : {
                    columnNumber : 4
                    functionName : executor
                    lineNumber : 12
                    scriptId : <scriptId>
                    url : test.js
                }
                [1] : {
                    columnNumber : 9
                    functionName : testConsoleTrace
                    lineNumber : 11
                    scriptId : <scriptId>
                    url : test.js
                }
            ]
            description : setTimeout
        }
    }
    timestamp : <timestamp>
    type : trace
}
Test with max size 0.
{
    columnNumber : 4
    exception : {
        className : Error
        description : Error     at bar (test.js:19:11)     at foo (test.js:23:5)     at testThrow (test.js:26:3)     at <anonymous>:1:1
        objectId : <objectId>
        subtype : error
        type : object
    }
    exceptionId : <exceptionId>
    lineNumber : 18
    scriptId : <scriptId>
    text : Uncaught
}
Test with max size 1.
{
    columnNumber : 4
    exception : {
        className : Error
        description : Error     at bar (test.js:19:11)     at foo (test.js:23:5)     at testThrow (test.js:26:3)     at <anonymous>:1:1
        objectId : <objectId>
        subtype : error
        type : object
    }
    exceptionId : <exceptionId>
    lineNumber : 18
    scriptId : <scriptId>
    stackTrace : {
        callFrames : [
            [0] : {
                columnNumber : 10
                functionName : bar
                lineNumber : 18
                scriptId : <scriptId>
                url : test.js
            }
        ]
    }
    text : Uncaught
}
Test with max size 2.
{
    columnNumber : 4
    exception : {
        className : Error
        description : Error     at bar (test.js:19:11)     at foo (test.js:23:5)     at testThrow (test.js:26:3)     at <anonymous>:1:1
        objectId : <objectId>
        subtype : error
        type : object
    }
    exceptionId : <exceptionId>
    lineNumber : 18
    scriptId : <scriptId>
    stackTrace : {
        callFrames : [
            [0] : {
                columnNumber : 10
                functionName : bar
                lineNumber : 18
                scriptId : <scriptId>
                url : test.js
            }
            [1] : {
                columnNumber : 4
                functionName : foo
                lineNumber : 22
                scriptId : <scriptId>
                url : test.js
            }
        ]
    }
    text : Uncaught
}
